feat: Add UserInfoService to retrieve username
This commit adds a new file `user-info.go` which contains the implementation of `UserInfoService`. This service has a method `GetUsername` that retrieves the username of a user given their ID. The implementation uses KeycloakClient to make requests to Keycloak API and returns an error if there is any issue with the request.