feat: Add delete user functionality to User Admin and User Edit controllers
This commit adds the ability to delete a user in both the User Admin and User Edit controllers. The `adminDeleteUser` function was added to the User Admin controller, while `deleteUser` was added to the User Edit controller. Both functions take a userID as input and call the `DeleteUser` method of their respective services. Additionally, routes were added for each function with appropriate Swagger documentation.